Ticket 333 - Testsuitesmanagement versioncheck needs a python interface
